requirements.txt
anyjson==0.3.3
celery==4.0.2
dj-database-url==0.3.0
Django==1.8.3
django-filter==0.10.0
django-grappelli==2.6.5
django-slack==5.8.0
djangorestframework==3.1.3
Markdown==2.6.2
opbeat==3.5.1
pika==0.10.0
psycopg2==2.6.1
pytz==2015.4
requests==2.7.0
tablib==0.10.0
uWSGI==2.0.15
